adding autorize for user manager and admins

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,19 +1,22 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
     public interface IRestaurantService
     {
-        void Update(int id, UpdateRestaurantDto dto);
-        void Delate(int id);
-        int Create(CreateRestaurantDto dto);
+        void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user);
+        void Delate(int id, ClaimsPrincipal user);
+        int Create(CreateRestaurantDto dto, int userId);
         IEnumerable<RestaurantDto> GetAll();
         RestaurantDto GetById(int id);
     }
@@ -23,31 +26,40 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger,
+            IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
-        public void Update(int id, UpdateRestaurantDto dto)
+        public void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user)
         {
-            var resaurant = _dbContext
+            var restaurant = _dbContext
                 .Restaurants
                 .FirstOrDefault(r => r.Id == id);
 
-            if (resaurant is null)
+            if (restaurant  is null)
                 throw new NotFoundException("Restaurant not found");
 
-            resaurant.Name = dto.Name;
-            resaurant.HasDelivery = dto.HasDelivery;
-            resaurant.Description = dto.Description;
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+                throw new ForbidException();
+
+            restaurant.Name = dto.Name;
+            restaurant.HasDelivery = dto.HasDelivery;
+            restaurant.Description = dto.Description;
             
             _dbContext.SaveChanges();
         }
 
-        public void Delate(int id)
+        public void Delate(int id, ClaimsPrincipal user)
         {
             _logger.LogError($"Restaurant with id {id} DELATE action invoked");
 
@@ -57,13 +69,20 @@ namespace RestaurantAPI.Services
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
 
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Delate)).Result;
+
+            if (!authorizationResult.Succeeded)
+                throw new ForbidException();
+
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
         }
 
-        public int Create(CreateRestaurantDto dto)
+        public int Create(CreateRestaurantDto dto, int userId)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = userId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
